// src/app/modules/dashboard/model/edit-profile.ts
export interface editProfile {
    name: string | null | undefined;
    email: string | null | undefined;
    password: string | null | undefined;
    confirmPassword: string | null | undefined;
    city: string | null | undefined;
    place: string | null | undefined;
    phoneNumber: string | null | undefined;
}